+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- examples/mongo: added [fosite-example](./examples/mongo) featuring mongo 
+  integration.
+
 ### Changed:
 - deps: updates to `fosite@v0.30.2`.
+- readme: references `examples/mongo` instead of having a wad of example code 
+  in the readme. 
 
 ## [v0.18.9] - 2020-06-13
 ### Fixed